refactor: Layers, Components 디렉토리 구조화 [#98]
- Header, Profile 파일을 Commo 에서 Layers, Components 로 이동
